chore(benchmark): refactor
